nightlyScale/requirements.txt
google-cloud-firestore
google-cloud-container
